@@ -11,37 +11,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.3",
-    "@angular/common": "~12.2.3",
-    "@angular/compiler": "~12.2.3",
-    "@angular/core": "~12.2.3",
-    "@angular/forms": "~12.2.3",
-    "@angular/localize": "~12.2.3",
-    "@angular/platform-browser": "~12.2.3",
-    "@angular/platform-browser-dynamic": "~12.2.3",
-    "@angular/router": "~12.2.3",
-    "ngx-virtual-scroller": "^4.0.3",
-    "rxjs": "~6.6.0",
-    "tslib": "^1.10.0",
+    "@angular/animations": "~13.0.0",
+    "@angular/cdk": "^13.0.0",
+    "@angular/common": "~13.0.0",
+    "@angular/compiler": "~13.0.0",
+    "@angular/core": "~13.0.0",
+    "@angular/forms": "~13.0.0",
+    "@angular/platform-browser": "~13.0.0",
+    "@angular/platform-browser-dynamic": "~13.0.0",
+    "@angular/router": "~13.0.0",
+    "rxjs": "~7.4.0",
+    "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.3",
-    "@angular/cli": "~12.2.3",
-    "@angular/compiler-cli": "~12.2.3",
-    "@types/jasmine": "~3.6.0",
+    "@angular-devkit/build-angular": "~13.0.0",
+    "@angular/cli": "~13.0.1",
+    "@angular/compiler-cli": "~13.0.0",
+    "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.4",
+    "jasmine-core": "~3.10.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.3.5"
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "typescript": "~4.4.3"
   }
 }
